+++ examples/smoke_tests/src/memory_sanity/memory_sanity.c
@@ -79,7 +79,9 @@ uint8_t write_readback_byte(uint32_t addr, uint8_t value, char verbose){
   return result;
 }
 
-int main() {  
+int main() {
+  // set peripherals to half freq
+  write_reg_u8(0x00030500, 0x2);
   init_uart(100000000/2, 3000000/2); // 50 MHz for simulation, 30 MHz for FPGA
   print_uart("[UART] Starting memory_sanity test\n");