+++ examples/smoke_tests/src/uart_sanity/uart_sanity.c
@@ -5,7 +5,10 @@
 #define OUTPUT_REG_ADDR    0x00030008
 #define TIMER_BASE_ADDR    0x00030200
 
-int main() {  
+int main() {
+  // set peripherals to half freq
+  write_reg_u8(0x00030500, 0x2);
+  
   init_uart(100000000/2, 3000000/2); // 50 MHz for simulation, 30 MHz for FPGA
   volatile int test = *(int*)(0x7000);
   print_uart("[UART] Hello from UART!\n");